refactor: split large command files into subfolders
streams.ts (2042 lines), zsets.ts (1018 lines), and lists.ts (1003 lines)
each split into a dedicated subfolder with one file per logical concern.
Public exports and listsCommands/zsetsCommands/streamsCommands arrays
preserved so commands/index.ts requires no changes.
