+++ src/commands/zsets/index.ts
@@ -1,0 +1,65 @@
+import { zaddCommand } from './zadd'
+import { zcardCommand, zremCommand } from './zrem'
+import { zrankCommand, zrevrankCommand } from './zrank'
+import { zmscoreCommand, zscoreCommand } from './zscore'
+import { zrandmemberCommand } from './zrandmember'
+import { zincrbyCommand } from './zincrby'
+import { zrangeCommand, zrevrangeCommand } from './zrange'
+import {
+  zcountCommand,
+  zrangebyscoreCommand,
+  zremrangebyscoreCommand,
+} from './zrangebyscore'
+import {
+  zlexcountCommand,
+  zrangebylexCommand,
+  zremrangebylexCommand,
+  zrevrangebylexCommand,
+} from './zrangebylex'
+import { zpopmaxCommand, zpopminCommand } from './zpop'
+
+export const zsetsCommands = [
+  zaddCommand,
+  zremCommand,
+  zcardCommand,
+  zrankCommand,
+  zrevrankCommand,
+  zscoreCommand,
+  zmscoreCommand,
+  zrandmemberCommand,
+  zincrbyCommand,
+  zrangeCommand,
+  zrevrangeCommand,
+  zrangebyscoreCommand,
+  zremrangebyscoreCommand,
+  zcountCommand,
+  zrangebylexCommand,
+  zrevrangebylexCommand,
+  zlexcountCommand,
+  zremrangebylexCommand,
+  zpopminCommand,
+  zpopmaxCommand,
+]
+
+export {
+  zaddCommand,
+  zremCommand,
+  zcardCommand,
+  zrankCommand,
+  zrevrankCommand,
+  zscoreCommand,
+  zmscoreCommand,
+  zrandmemberCommand,
+  zincrbyCommand,
+  zrangeCommand,
+  zrevrangeCommand,
+  zrangebyscoreCommand,
+  zremrangebyscoreCommand,
+  zcountCommand,
+  zrangebylexCommand,
+  zrevrangebylexCommand,
+  zlexcountCommand,
+  zremrangebylexCommand,
+  zpopminCommand,
+  zpopmaxCommand,
+}
